XD-3412: Refactoring
unknownHostsExpression

Pull up common properties to AbstractRemoteServerProperties.

Move test server config to test server.
