+++ common/spring-cloud-stream-modules-common-configuration/src/main/java/org/springframework/cloud/stream/module/sftp/SftpSessionFactoryConfiguration.java
@@ -40,8 +40,8 @@ public class SftpSessionFactoryConfiguration {
 		sftpSessionFactory.setUser(this.config.getUsername());
 		sftpSessionFactory.setPassword(this.config.getPassword());
 		sftpSessionFactory.setAllowUnknownKeys(this.config.isAllowUnknownKeys());
-		if (!SftpSessionFactoryProperties.KNOWNHOSTS_UNDEFINED.equals(this.config.getKnownHosts())) {
-			sftpSessionFactory.setKnownHosts("#{" + this.config.getKnownHosts() + "}");
+		if (!SftpSessionFactoryProperties.KNOWNHOSTS_UNDEFINED.equals(this.config.getKnownHostsExpression())) {
+			sftpSessionFactory.setKnownHosts("#{" + this.config.getKnownHostsExpression() + "}");
 		}
 		return sftpSessionFactory;
 	}
