+++ common/spring-cloud-stream-modules-test-support/src/main/java/org/springframework/cloud/stream/modules/test/ftp/TestSftpServer.java
@@ -16,6 +16,8 @@
 package org.springframework.cloud.stream.modules.test.ftp;
 
 import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
 import java.util.Arrays;
 
 import javax.annotation.PostConstruct;
@@ -30,6 +32,7 @@ import org.apache.sshd.server.Command;
 import org.apache.sshd.server.PasswordAuthenticator;
 import org.apache.sshd.server.keyprovider.SimpleGeneratorHostKeyProvider;
 import org.apache.sshd.server.sftp.SftpSubsystem;
+import org.junit.rules.TemporaryFolder;
 
 /**
  * @author David Turanski
@@ -40,6 +43,59 @@ public class TestSftpServer extends TestFtpServer {
 
 	public TestSftpServer(String root, int port) {
 		super(root);
+		setFtpTemporaryFolder(
+				new TemporaryFolder() {
+					@Override
+					public void create() throws IOException {
+						super.create();
+						File rootFolder = this.newFolder(getRootFolderName());
+						File sourceFtpDirectory = new File(rootFolder, "sftpSource");
+						sourceFtpDirectory.mkdir();
+
+						File file = new File(sourceFtpDirectory, "sftpSource1.txt");
+						file.createNewFile();
+						FileOutputStream fos = new FileOutputStream(file);
+						fos.write("source1".getBytes());
+						fos.close();
+						file = new File(sourceFtpDirectory, "sftpSource2.txt");
+						file.createNewFile();
+						fos = new FileOutputStream(file);
+						fos.write("source2".getBytes());
+						fos.close();
+
+						File targetFtpDirectory = new File(rootFolder, "sftpTarget");
+						targetFtpDirectory.mkdir();
+
+						setFtpRootFolder(rootFolder);
+						setSourceFtpDirectory(sourceFtpDirectory);
+						setTargetFtpDirectory(targetFtpDirectory);
+					}
+				});
+		setLocalTemporaryFolder(new TemporaryFolder() {
+
+			@Override
+			public void create() throws IOException {
+				super.create();
+				File rootFolder = getRootFolder();
+				File sourceLocalDirectory = new File(rootFolder, "localSource");
+				sourceLocalDirectory.mkdirs();
+				File file = new File(sourceLocalDirectory, "localSource1.txt");
+				file.createNewFile();
+				file = new File(sourceLocalDirectory, "localSource2.txt");
+				file.createNewFile();
+
+				File subSourceLocalDirectory = new File(sourceLocalDirectory, "subLocalSource");
+				subSourceLocalDirectory.mkdir();
+				file = new File(subSourceLocalDirectory, "subLocalSource1.txt");
+				file.createNewFile();
+
+				File targetLocalDirectory = new File(rootFolder, "slocalTarget");
+				targetLocalDirectory.mkdir();
+
+				setSourceLocalDirectory(sourceLocalDirectory);
+				setTargetLocalDirectory(targetLocalDirectory);
+			}
+		});
 	}
 
 	@Override
